Refactor the frontend.
When it was time to do the 2.1 API, I figured it'd be good to simply
copy and paste the entire frontend to minimize compatibility issues.
That doesn't really work because a good deal of code (eg the database)
is shared, so really everything in one API has to be sensitive to the
overall system. Also, a great deal of delicate code (amb2.py, for
example), winds up duplicated with very little changes.

This patch splits the API-specific stuff out of generators, so we can
reduce the number of generators from 4 to 2. The API specific parts have
been renamed to what they actually are: ContextManagers. Some of that
can be shared and now exists in a base class, but most of the details of
context management will remain API specific.

An amount of the C++-specific stuff has been moved out of generators as
well, since this tended to be API specific.

While we're at it, this also simplifies the directory structure a bit.
