+++ ambuild2/frontend/v2_0/amb2.py
@@ -1,4 +1,4 @@
-# vim: set ts=8 sts=2 sw=2 tw=99 et:
+# vim: set ts=8 sts=4 sw=4 tw=99 et:
 #
 # This file is part of AMBuild.
 #
@@ -14,3 +14,19 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with AMBuild. If not, see <http://www.gnu.org/licenses/>.
+import os
+from ambuild2 import util
+from ambuild2.frontend import amb2
+from ambuild2.frontend.v2_0.cpp import detect
+
+class Generator(amb2.Generator):
+    def __init__(self, cm):
+        super(Generator, self).__init__(cm)
+
+    def detectCompilers(self):
+        if not self.compiler:
+            with util.FolderChanger(self.cacheFolder):
+                self.base_compiler = detect.DetectCxx(os.environ, self.cm.options)
+                self.compiler = self.base_compiler.clone()
+
+        return self.compiler
