--- ambuild2/frontend/v2_0/base/__init__.py
@@ -1,1 +1,0 @@
-from ambuild2.frontend.v2_0.base.gen import BaseGenerator
