--- ambuild2/frontend/v2_1/amb2/__init__.py
@@ -1,1 +1,0 @@
-from ambuild2.frontend.v2_1.amb2.gen import Generator
