--- ambuild2/frontend/v2_1/base/__init__.py
@@ -1,1 +1,0 @@
-from ambuild2.frontend.v2_1.base.gen import BaseGenerator
