+++ ambuild2/frontend/v2_1/context.py
@@ -30,9 +30,10 @@ from ambuild2.frontend.version import Version
 # scripts change.
 
 class BaseContext(object):
-    def __init__(self, generator, parent, vars, script):
+    def __init__(self, cm, parent, vars, script):
         super(BaseContext, self).__init__()
-        self.generator_ = generator
+        self.cm = cm
+        self.generator_ = cm.generator
         self.parent_ = parent
         self.script_ = script
 
@@ -53,23 +54,23 @@ class BaseContext(object):
     # Root source folder.
     @property
     def sourcePath(self):
-        return self.generator_.sourcePath
+        return self.cm.sourcePath
 
     @property
     def options(self):
-        return self.generator_.options
+        return self.cm.options
 
     @property
     def target(self):
-        return self.generator_.target
+        return self.cm.target
 
     @property
     def host(self):
-        return self.generator_.host
+        return self.cm.host
 
     @property
     def originalCwd(self):
-        return self.generator_.originalCwd
+        return self.cm.originalCwd
 
     @property
     def backend(self):
@@ -77,17 +78,17 @@ class BaseContext(object):
 
     @property
     def buildPath(self):
-        return self.generator_.buildPath
+        return self.cm.buildPath
 
     @property
     def apiVersion(self):
         return Version('2.1.1')
 
     def Import(self, path, vars = None):
-        return self.generator_.importScript(self, path, vars or {})
+        return self.cm.importScript(self, path, vars or {})
 
     def Eval(self, path, vars = None):
-        return self.generator_.evalScript(self, path, vars or {})
+        return self.cm.evalScript(self, path, vars or {})
 
     def AddConfigureFile(self, path):
         return self.generator_.addConfigureFile(self, path)
@@ -99,14 +100,11 @@ class EmptyContext(BaseContext):
 
 # Access to input- and output-oriented API.
 class BuildContext(BaseContext):
-    # This nonce is an input flag to AddCommand.
-    ALWAYS_DIRTY = object()
-
     # Provide an accessor so users don't have to import the v2_1 namespace.
     tools = tools
 
-    def __init__(self, generator, parent, vars, script, sourceFolder, buildFolder):
-        super(BuildContext, self).__init__(generator, parent, vars, script)
+    def __init__(self, cm, parent, vars, script, sourceFolder, buildFolder):
+        super(BuildContext, self).__init__(cm, parent, vars, script)
         self.localFolder_ = None
         self.cxx_ = None
 
@@ -115,7 +113,7 @@ class BuildContext(BaseContext):
 
         self.sourceFolder = sourceFolder
         self.buildFolder = buildFolder
-        self.currentSourcePath = os.path.join(generator.sourcePath, sourceFolder)
+        self.currentSourcePath = os.path.join(cm.sourcePath, sourceFolder)
         self.currentSourceFolder = sourceFolder
         self.buildFolder = buildFolder
 
@@ -127,7 +125,7 @@ class BuildContext(BaseContext):
             self.buildFolder = os.path.normpath(self.buildFolder)
 
     def Build(self, path, vars = None):
-        return self.generator_.runBuildScript(self, path, vars or {})
+        return self.cm.runBuildScript(self, path, vars or {})
 
     # Any consumed options will be removed from the given dictionary. Callers
     # can detect unconsumed options by inspecting the dictionary after.
@@ -140,6 +138,10 @@ class BuildContext(BaseContext):
     @property
     def cxx(self):
         return self.cxx_
+
+    @property
+    def ALWAYS_DIRTY(self):
+        return self.cm.ALWAYS_DIRTY
 
     # In build systems with dependency graphs, this can return a node
     # representing buildFolder. Otherwise, it returns buildFolder.
@@ -181,7 +183,7 @@ class BuildContext(BaseContext):
                                                env_data = env_data)
 
     def Context(self, name):
-        return self.generator_.Context(name)
+        return self.cm.Context(name)
 
     def Add(self, taskbuilder):
         taskbuilder.finish(self)
@@ -199,14 +201,14 @@ class BuildContext(BaseContext):
 
 # Access to everything.
 class TopLevelBuildContext(BuildContext):
-    def __init__(self, generator, parent, vars, script, sourceFolder, buildFolder):
-        super(TopLevelBuildContext, self).__init__(generator, parent, vars, script, sourceFolder,
+    def __init__(self, cm, parent, vars, script, sourceFolder, buildFolder):
+        super(TopLevelBuildContext, self).__init__(cm, parent, vars, script, sourceFolder,
                                                    buildFolder)
 
 # The root build context.
 class RootBuildContext(BuildContext):
-    def __init__(self, generator, vars, script):
-        super(RootBuildContext, self).__init__(generator = generator,
+    def __init__(self, cm, vars, script):
+        super(RootBuildContext, self).__init__(cm = cm,
                                                parent = None,
                                                vars = vars,
                                                script = script,
