+++ ambuild2/frontend/v2_1/context_manager.py
@@ -1,4 +1,4 @@
-# vim: set ts=8 sts=2 sw=2 tw=99 et:
+# vim: set ts=8 sts=4 sw=4 tw=99 et:
 #
 # This file is part of AMBuild.
 #
@@ -14,31 +14,27 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with AMBuild. If not, see <http://www.gnu.org/licenses/>.
+import copy
 import os
-import sys, copy
+import sys
 from ambuild2 import util
+from ambuild2.frontend import context_manager
 from ambuild2.frontend import paths
 from ambuild2.frontend.system import System
-from ambuild2.frontend.v2_1.base.context import \
-    TopLevelBuildContext, \
-    BuildContext, \
-    EmptyContext, \
-    RootBuildContext
+from ambuild2.frontend.v2_1.context import BuildContext
+from ambuild2.frontend.v2_1.context import EmptyContext
+from ambuild2.frontend.v2_1.context import RootBuildContext
+from ambuild2.frontend.v2_1.context import TopLevelBuildContext
+from ambuild2.frontend.v2_1.cpp import detect
+from ambuild2.frontend.version import Version
 
 class ConfigureException(Exception):
     def __init__(self, *args, **kwargs):
         super(ConfigureException, self).__init__(*args, **kwargs)
 
-class BaseGenerator(object):
+class ContextManager(context_manager.ContextManager):
     def __init__(self, sourcePath, buildPath, originalCwd, options, args):
-        super(BaseGenerator, self).__init__()
-        self.sourcePath = sourcePath
-        self.buildPath = os.path.normpath(buildPath)
-        self.originalCwd = originalCwd
-        self.options = options
-        self.args = args
-        self.contextStack_ = []
-        self.configure_failed = False
+        super(ContextManager, self).__init__(sourcePath, buildPath, originalCwd, options, args)
 
         # Detect the target architecture.
         self.host = System.Host
@@ -49,18 +45,16 @@ class BaseGenerator(object):
         if new_arch is not None:
             self.target = System(self.target.platform, util.NormalizeArchString(new_arch))
 
+    @property
+    def apiVersion(self):
+        return Version('2.1.1')
+
     def parseBuildScripts(self):
         root = os.path.join(self.sourcePath, 'AMBuildScript')
-        self.addConfigureFile(None, root)
+        self.generator.addConfigureFile(None, root)
 
         cx = RootBuildContext(self, {}, root)
         self.execContext(cx)
-
-    def pushContext(self, cx):
-        self.contextStack_.append(cx)
-
-    def popContext(self):
-        self.contextStack_.pop()
 
     def importScript(self, context, path, vars = None):
         if not isinstance(path, util.StringType()):
@@ -110,7 +104,7 @@ class BaseGenerator(object):
 
         # Get the absolute script path.
         scriptPath = os.path.join(self.sourcePath, scriptFile)
-        self.addConfigureFile(parent, scriptPath)
+        self.generator.addConfigureFile(parent, scriptPath)
 
         # Make the new context. We allow top-level contexts in the root build
         # and otherwise for absolute paths.
@@ -122,7 +116,7 @@ class BaseGenerator(object):
                 raise Exception("Nested build contexts must be within the same folder structure")
             constructor = BuildContext
 
-        cx = constructor(generator = self,
+        cx = constructor(cm = self,
                          parent = parent,
                          vars = vars,
                          script = scriptPath,
@@ -148,16 +142,6 @@ class BaseGenerator(object):
 
         return scriptGlobals
 
-    def compileScript(self, path):
-        with open(path) as fp:
-            chars = fp.read()
-
-            # Python 2.6 can't compile() with Windows line endings?!?!!?
-            chars = chars.replace('\r\n', '\n')
-            chars = chars.replace('\r', '\n')
-
-            return compile(chars, path, 'exec')
-
     def computeScriptPaths(self, parent, target):
         # By default, all generated files for a build script are placed in a path
         # matching its layout in the source tree.
@@ -182,60 +166,22 @@ class BaseGenerator(object):
 
         return sourceFolder, buildFolder, os.path.join(sourceFolder, name)
 
-    def generateBuildFiles(self):
-        build_py = os.path.join(self.buildPath, 'build.py')
-        with open(build_py, 'w') as fp:
-            fp.write("""
-#!{exe}
-# vim set: ts=8 sts=2 sw=2 tw=99 et:
-import sys
-from ambuild2 import run
-
-if not run.CompatBuild(r"{build}"):
-  sys.exit(1)
-""".format(exe = sys.executable, build = self.buildPath))
-
-        with open(os.path.join(self.buildPath, 'Makefile'), 'w') as fp:
-            fp.write("""
-all:
-	"{exe}" "{py}"
-""".format(exe = sys.executable, py = build_py))
-
-    def generate(self):
-        self.preGenerate()
-        self.parseBuildScripts()
-        self.postGenerate()
-        if self.options.make_scripts:
-            self.generateBuildFiles()
-        return True
-
     def getLocalFolder(self, context):
         return context.buildFolder
 
-    @property
-    def backend(self):
-        raise Exception('Must be implemented!')
+    def copyCompilerVars(self, vars, compiler):
+        for prop_name in compiler.vendor.extra_props:
+            key = '{0}_{1}'.format(compiler.vendor.name, prop_name)
+            vars[key] = compiler.vendor.extra_props[prop_name]
 
-    def addSymlink(self, context, source, output_path):
-        raise Exception('Must be implemented!')
-
-    def addFolder(self, context, folder):
-        raise Exception('Must be implemented!')
-
-    def addCopy(self, context, source, output_path):
-        raise Exception('Must be implemented!')
-
-    def addShellCommand(self,
-                        context,
-                        inputs,
-                        argv,
-                        outputs,
-                        folder = -1,
-                        dep_type = None,
-                        weak_inputs = [],
-                        shared_outputs = [],
-                        env_data = None):
-        raise Exception('Must be implemented!')
-
-    def addConfigureFile(self, context, path):
-        raise Exception('Must be implemented!')
+    def createGenerator(self, name):
+        if name == 'vs':
+            from ambuild2.frontend.v2_1.vs.gen import Generator
+            self.generator = Generator(self)
+        elif name == 'ambuild2':
+            # Different name, because pip, impressively, does not delete caches for old files, and
+            # and this causes the old "amb2" directory to linger and create import problems!
+            from ambuild2.frontend.v2_1.amb2_gen import Generator
+            self.generator = Generator(self)
+        else:
+            super(ContextManager, self).createGenerator(name)
