+++ ambuild2/frontend/v2_1/prep.py
@@ -151,19 +151,13 @@ class Preparer(object):
                 os.mkdir(new_buildpath)
             self.buildPath = new_buildpath
 
-        if options.generator == 'ambuild2':
-            from ambuild2.frontend.v2_1.amb2 import gen
-            builder = gen.Generator(self.sourcePath, self.buildPath, os.getcwd(), options, args)
-        elif options.generator == 'vs':
-            from ambuild2.frontend.v2_1.vs import gen
-            builder = gen.Generator(self.sourcePath, self.buildPath, os.getcwd(), options, args)
-        else:
-            sys.stderr.write('Unrecognized build generator: ' + options.generator + '\n')
-            sys.exit(1)
+        from ambuild2.frontend.v2_1.context_manager import ContextManager
+
+        cm = ContextManager(self.sourcePath, self.buildPath, os.getcwd(), options, args)
 
         with util.FolderChanger(self.buildPath):
             try:
-                if not builder.generate():
+                if not cm.generate(options.generator):
                     sys.stderr.write('Configure failed.\n')
                     sys.exit(1)
             except Exception as e:
