+++ ambuild2/frontend/v2_1/vs/gen.py
@@ -14,86 +14,12 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with AMBuild. If not, see <http://www.gnu.org/licenses/>.
-import os, errno
-import uuid as uuids
-from ambuild2 import util
-from ambuild2 import nodetypes
-from ambuild2.frontend import paths
+import ambuild2.frontend.vs.gen as vs_gen
 from ambuild2.frontend.v2_1.vs import cxx
-from ambuild2.frontend.v2_1.vs import nodes
-from ambuild2.frontend.v2_1.base import BaseGenerator
 
-SupportedVersions = ['10', '11', '12', '14', '15', '16']
-YearMap = {
-    '2010': 10,
-    '2012': 11,
-    '2013': 12,
-    '2015': 14,
-    '2017': 15,
-    '2019': 16,
-}
-
-class Generator(BaseGenerator):
-    def __init__(self, sourcePath, buildPath, originalCwd, options, args):
-        super(Generator, self).__init__(sourcePath, buildPath, originalCwd, options, args)
-        self.compiler = None
-        self.vs_version = None
-        self.files_ = {}
-        self.projects_ = set()
-
-        if self.options.vs_version in SupportedVersions:
-            self.vs_version = int(self.options.vs_version)
-        else:
-            if self.options.vs_version not in YearMap:
-                util.con_err(
-                    util.ConsoleRed,
-                    'Unsupported Visual Studio version: {0}'.format(self.options.vs_version),
-                    util.ConsoleNormal)
-                raise Exception('Unsupported Visual Studio version: {0}'.format(
-                    self.options.vs_version))
-            self.vs_version = YearMap[self.options.vs_version]
-
-        self.cacheFile = os.path.join(self.buildPath, '.cache')
-        try:
-            with open(self.cacheFile, 'rb') as fp:
-                self.vars_ = util.pickle.load(fp)
-        except:
-            self.vars_ = {}
-
-        if 'uuids' not in self.vars_:
-            self.vars_['uuids'] = {}
-
-        self.target_platform = 'windows'
-
-    # Overridden.
-    @property
-    def backend(self):
-        return 'vs'
-
-    # Overridden.
-    def preGenerate(self):
-        pass
-
-    # Overriden.
-    def postGenerate(self):
-        self.generateProjects()
-        with open(self.cacheFile, 'wb') as fp:
-            util.DiskPickle(self.vars_, fp)
-
-    def generateProjects(self):
-        for node in self.projects_:
-            # We cache uuids across runs to keep them consistent.
-            node.uuid = self.vars_['uuids'].get(node.path)
-            if node.uuid is None:
-                node.uuid = str(uuids.uuid1()).upper()
-                self.vars_['uuids'][node.path] = node.uuid
-            node.project.export(node)
-
-    # Overridden.
-    #
-    # We don't support reconfiguring in this frontend.
-    def addConfigureFile(self, cx, path):
-        pass
+class Generator(vs_gen.Generator):
+    def __init__(self, cm):
+        super(Generator, self).__init__(cm)
 
     # Overridden.
     def detectCompilers(self, options):
@@ -103,78 +29,3 @@ class Generator(BaseGenerator):
             self.base_compiler = cxx.Compiler(vendor)
             self.compiler = self.base_compiler.clone()
         return self.compiler
-
-    # Overridden.
-    def enterContext(self, cx):
-        cx.vs_nodes = []
-
-    # Overridden.
-    def leaveContext(self, cx):
-        pass
-
-    def ensureUnique(self, path):
-        if path in self.files_:
-            entry = self.files_[path]
-            util.con_err(util.ConsoleRed,
-                         'Path {0} already exists as: {1}'.format(path,
-                                                                  entry.kind), util.ConsoleNormal)
-            raise Exception('Path {0} already exists as: {1}'.format(path, entry.kind))
-
-    # Overridden.
-    def getLocalFolder(self, context):
-        if type(context.localFolder_) is nodes.FolderNode or context.localFolder_ is None:
-            return context.localFolder_
-
-        if not len(context.buildFolder):
-            context.localFolder_ = None
-        else:
-            context.localFolder_ = self.addFolder(context.parent, context.buildFolder)
-
-        return context.localFolder_
-
-    # Overridden.
-    def addFolder(self, cx, folder):
-        parentFolderNode = None
-        if cx is not None:
-            parentFolderNode = cx.localFolder
-
-        _, path = paths.ResolveFolder(parentFolderNode, folder)
-        if path in self.files_:
-            entry = self.files_[path]
-            if type(entry) is not nodes.FolderNode:
-                self.ensureUnique(path)  # Will always throw.
-            return entry
-
-        try:
-            os.makedirs(path)
-        except OSError as exn:
-            if not (exn.errno == errno.EEXIST and os.path.isdir(path)):
-                raise
-
-        obj = nodes.FolderNode(path)
-        self.files_[path] = obj
-        return obj
-
-    # Overridden.
-    def addShellCommand(self,
-                        context,
-                        inputs,
-                        argv,
-                        outputs,
-                        folder = -1,
-                        dep_type = None,
-                        weak_inputs = None,
-                        shared_outputs = None):
-        print(inputs, argv, outputs, folder, dep_type, weak_inputs, shared_outputs)
-
-    def addOutput(self, context, path, parent):
-        self.ensureUnique(path)
-
-        node = nodes.OutputNode(context, path, parent)
-        self.files_[path] = node
-        return node
-
-    def addProjectNode(self, context, project):
-        self.ensureUnique(project.path)
-        self.projects_.add(project)
-        self.files_[project.path] = project
