+++ ambuild2/frontend/vs/xmlbuilder.py
